Evaluate the simulated users
