+++ .github/workflows/azure-dev.yml
@@ -112,6 +112,7 @@ jobs:
       AZURE_CONTAINER_APPS_WORKLOAD_PROFILE: ${{ vars.AZURE_CONTAINER_APPS_WORKLOAD_PROFILE }}
       USE_CHAT_HISTORY_BROWSER: ${{ vars.USE_CHAT_HISTORY_BROWSER }}
       USE_MEDIA_DESCRIBER_AZURE_CU: ${{ vars.USE_MEDIA_DESCRIBER_AZURE_CU }}
+      USE_AI_PROJECT: ${{ vars.USE_AI_PROJECT }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
