+++ azure.yaml
@@ -124,6 +124,7 @@ pipeline:
       - AZURE_CONTAINER_APPS_WORKLOAD_PROFILE
       - USE_CHAT_HISTORY_BROWSER
       - USE_MEDIA_DESCRIBER_AZURE_CU
+      - USE_AI_PROJECT
   secrets:
       - AZURE_SERVER_APP_SECRET
       - AZURE_CLIENT_APP_SECRET
