+++ evals/safety_results.json
@@ -1,0 +1,26 @@
+{
+  "hate_unfairness": {
+    "low_count": 20,
+    "score_total": 8,
+    "mean_score": 0.4,
+    "low_rate": 1.0
+  },
+  "sexual": {
+    "low_count": 20,
+    "score_total": 9,
+    "mean_score": 0.45,
+    "low_rate": 1.0
+  },
+  "violence": {
+    "low_count": 20,
+    "score_total": 9,
+    "mean_score": 0.45,
+    "low_rate": 1.0
+  },
+  "self_harm": {
+    "low_count": 20,
+    "score_total": 10,
+    "mean_score": 0.5,
+    "low_rate": 1.0
+  }
+}
