--- infra/core/security/keyvault-access.bicep
@@ -1,22 +1,0 @@
-metadata description = 'Assigns an Azure Key Vault access policy.'
-param name string = 'add'
-
-param keyVaultName string
-param permissions object = { secrets: [ 'get', 'list' ] }
-param principalId string
-
-resource keyVaultAccessPolicies 'Microsoft.KeyVault/vaults/accessPolicies@2022-07-01' = {
-  parent: keyVault
-  name: name
-  properties: {
-    accessPolicies: [ {
-        objectId: principalId
-        tenantId: subscription().tenantId
-        permissions: permissions
-      } ]
-  }
-}
-
-resource keyVault 'Microsoft.KeyVault/vaults@2022-07-01' existing = {
-  name: keyVaultName
-}
